requirements.txt
torch==0.4.1
torchvision==0.2.1
numpy==1.15.1
Pillow=5.2.0

